use size_t for length
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@14619 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/roken/hex.c
@@ -54,7 +54,7 @@ pos(char c)
 }
 
 ssize_t
-hex_encode(const void *data, int size, char **str)
+hex_encode(const void *data, size_t size, char **str)
 {
     const unsigned char *q = data;
     size_t i;
